pinephone: make systemd-boot the installed image bootloader
This change requires an UEFI capable SPL to be installed, like tow-boot.
The boot chain now becomes:
  1. SPL
  2. eMMC boot => Tow-Boot
  3. systemd-boot EFI bootloader
  4. through default menu choice, will boot LuneOS entry

Notice that if a SDCard with U-Boot bootloader is present, it will always
be booted before Tow-Boot.

Notice 2: currently there is no way to start the LuneOS recovery UI using
this boot chain, as the Volume keys interactions are already used by
Tow-Boot.

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>