pinephone wic boot.conf: specificy which input device triggers recovery
Now that Tow-boot is used on Pinephone, we can't have different kernel
cmdline depending on volume key press at early boot.
There is also no UI to selct which EFI image to boot currently.

However, it is now possible to test if a volume key is pressed after
systemd-boot has been started by Tow-boot, thanks to the new key-state
utility.
So use it to trigger the start of luneos-recoveryui is a Volume key is
pressed.

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>